Unit tests for MySqli connection class.
These require a database in order to run, so are not true unit tests, but at least do test the code. True unit tests cannot be written for this at the moment, due to the use of public attributes on the mysqli class built into PHP. A possible solution may be to change the connection class to use an extension of the PHP mysqli class, and on the extension define accessor methods that can be mocked for accessing the public properties.